Disable color scheme to decrease FF boot time
The solarized-dark color scheme
(https://github.com/claytron/pentadactyl-solarized) was causing a very
long boot up/window open time for Firefox. I use my window manager's
tabs instead of Firefox's built in tabs. Because of this, the cost of a
slow startup time more than outweighs the benefit of having a nice color
scheme. So until I figure out if it is possible to reduce this startup
time, I'm opting to disable the color scheme.