Remove dependency __git_ps1 function
The official git repository contains the file
contrib/completion/git-prompt.sh which defines the __git_ps1 function.
It seems that this file did not come with my git installation and even
if it did, git can be installed in any number of locations. This makes
it difficult to programmatically determine its location (in order to
source it). So instead of depending on __git_ps1, it is simpler for me
to just write a git_head function which outputs the branch name
without relying on any external dependencies.